+++ src/elements/PaymentForm.php
@@ -634,7 +634,7 @@ class PaymentForm extends Element
 
         $applyTax = false;
         // Tax logic - apply just to subscriptions
-        if ($this->enableSubscriptions ){
+        if ($this->enableSubscriptions){
             $applyTax = true;
         }
 
@@ -659,6 +659,7 @@ class PaymentForm extends Element
             'multiplePlansAmounts' => $multiplePlansAmounts,
             'setupFees' => $setupFees,
             'applyTax' => $applyTax,
+            'enableTaxes' => $this->settings->enableTaxes,
             'tax' => $this->settings->tax,
             'currencySymbol' => $this->getCurrencySymbol(),
             'taxLabel' => Craft::t('site', 'Tax Amount'),